chore: Refactor api/client models folder structure
