+++ api/src/models/api/stardust/foundry/IFoundriesResponse.ts
@@ -1,5 +1,5 @@
 import { IOutputsResponse } from "@iota/iota.js-stardust";
-import { IResponse } from "../IResponse";
+import { IResponse } from "../../IResponse";
 
 export interface IFoundriesResponse extends IResponse {
     /**
