+++ api/src/models/api/stardust/foundry/IFoundryRequest.ts
