+++ api/src/models/api/stardust/nft/INftAddressDetailsResponse.ts
@@ -1,5 +1,5 @@
 import { IOutputResponse } from "@iota/iota.js-stardust";
-import { IResponse } from "../IResponse";
+import { IResponse } from "../../IResponse";
 
 export interface INftAddressDetailsResponse extends IResponse {
     /**
