+++ api/src/models/api/stardust/nft/INftDetailsRequest.ts
