+++ api/src/models/api/stardust/nft/INftOutputsRequest.ts
