+++ api/src/models/api/stardust/nft/INftOutputsResponse.ts
@@ -1,5 +1,5 @@
 import { IOutputsResponse } from "@iota/iota.js-stardust";
-import { IResponse } from "../IResponse";
+import { IResponse } from "../../IResponse";
 
 export interface INftOutputsResponse extends IResponse {
     /**
