+++ api/src/routes/stardust/alias/foundries/get.ts
@@ -1,6 +1,6 @@
 import { ServiceFactory } from "../../../../factories/serviceFactory";
-import { IFoundriesRequest } from "../../../../models/api/stardust/IFoundriesRequest";
-import { IFoundriesResponse } from "../../../../models/api/stardust/IFoundriesResponse";
+import { IFoundriesRequest } from "../../../../models/api/stardust/foundry/IFoundriesRequest";
+import { IFoundriesResponse } from "../../../../models/api/stardust/foundry/IFoundriesResponse";
 import { IConfiguration } from "../../../../models/configuration/IConfiguration";
 import { STARDUST } from "../../../../models/db/protocolVersion";
 import { NetworkService } from "../../../../services/networkService";
