+++ api/src/routes/stardust/nft/outputs/get.ts
@@ -1,6 +1,6 @@
 import { ServiceFactory } from "../../../../factories/serviceFactory";
-import { INftOutputsRequest } from "../../../../models/api/stardust/INftOutputsRequest";
-import { INftOutputsResponse } from "../../../../models/api/stardust/INftOutputsResponse";
+import { INftOutputsRequest } from "../../../../models/api/stardust/nft/INftOutputsRequest";
+import { INftOutputsResponse } from "../../../../models/api/stardust/nft/INftOutputsResponse";
 import { IConfiguration } from "../../../../models/configuration/IConfiguration";
 import { STARDUST } from "../../../../models/db/protocolVersion";
 import { NetworkService } from "../../../../services/networkService";
